exercise.main/Basket.cs
using System.Collections.Generic;
using System.Diagnostics.Contracts;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace exercise.main
{
    public class Basket : Inventory

    {
        private int _capacity;
        private List<InventoryProducts> _items = new List<InventoryProducts>();

       

        public bool IsFull()
        {

            if (_capacity == 0)
            {
                return true;
            }
            else
            {
                return false;
            }
        }

        public double TotalCost()
        {
            double total = 0;
            foreach (var item in _items)
            {
                total += item.Price;
            }

            return total;
        }

        public List<InventoryProducts> Items { get { return _items; } }
        public int Capacity { get { return _capacity; } set { _capacity = value; } }
    }
}
